requirements.txt
numpy<1.17.0
tensorflow-gpu==1.14.0
gast==0.2.2